moving some files around & resolving eslint issues
