+++ src/pages/dashboard/map-container.scss
@@ -90,6 +90,9 @@ $death-color: #404856;
   flex-direction: row;
   align-items: center;
   padding-top: 10px;
+  label {
+    display: none;
+  }
   #map-start-stop {
     cursor: pointer;
     display: inline-block;
